Refactor sms module
