+++ packages/attestation-service/src/helper/anonymity.ts
@@ -1,0 +1,7 @@
+export function obfuscateNumber(phoneNumber: string): string {
+  try {
+    return phoneNumber.slice(0, 7) + '...'
+  } catch {
+    return ''
+  }
+}
