+++ packages/attestation-service/src/sms/provider/smsProvider.enum.ts
@@ -1,0 +1,7 @@
+export enum SmsProviderType {
+  NEXMO = 'nexmo',
+  UNKNOWN = 'unknown',
+  TWILIO = 'twilio',
+  MESSAGEBIRD = 'messagebird',
+  TELEKOM = 'telekom',
+}
