+++ packages/attestation-service/src/sms/requests/rerequestAttestation.request.ts
@@ -1,0 +1,12 @@
+import { AttestationKey } from '../../models/attestation'
+import Logger from 'bunyan'
+import { SequelizeLogger } from '../../db'
+
+export interface RerequestAttestationRequest {
+  key: AttestationKey
+  appSignature: string | undefined
+  language: string | undefined
+  securityCodePrefix: string
+  logger: Logger
+  sequelizeLogger: SequelizeLogger
+}
