+++ packages/attestation-service/src/sms/sms.interface.ts
@@ -1,0 +1,27 @@
+import { SmsProviderType } from './provider/smsProvider.enum'
+import { SmsProvider } from './provider/smsProvider'
+import { AttestationModel, AttestationStatus } from '../models/attestation'
+import Logger from 'bunyan'
+import { RerequestAttestationRequest } from './requests/rerequestAttestation.request'
+import { SendSmsRequest } from './requests/sendSms.request'
+
+export interface ISmsService {
+  initializeSmsProviders(deliveryStatusURLForProviderType: (type: string) => string): Promise<void>
+
+  startSendSms(smsRequest: SendSmsRequest): Promise<AttestationModel>
+  rerequestAttestation(
+    rerequestAttestationRequest: RerequestAttestationRequest
+  ): Promise<AttestationModel>
+
+  receivedDeliveryReport(
+    deliveryId: string,
+    deliveryStatus: AttestationStatus,
+    errorCode: string | null,
+    logger: Logger
+  ): Promise<void>
+
+  smsProviderOfType(type: SmsProviderType): SmsProvider | undefined
+  configuredSmsProviders(): string[]
+  smsProvidersWithDeliveryStatus(): SmsProvider[]
+  unsupportedRegionCodes(): string[]
+}
